fix(release): split permanent and historical consumer gates
